Correct LogFilter class and add log2.txt file to check .save method

+++ job4j_design/src/main/java/ru/job4j/io/LogFilter.java
@@ -1,38 +1,59 @@
 package ru.job4j.io;
 
+import static org.apache.commons.lang3.StringUtils.isBlank;
+
 import java.io.*;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class LogFilter {
+    private final static String DELIMITER_LINE = " ";
+
     public static List<String> filter(String file) {
-        List<String> lines = new ArrayList<>();
-        try (BufferedReader buffer = new BufferedReader(new FileReader(file))) {
-            buffer.lines()
-                    .filter(line -> {
-                        String[] words = line.split(" ");
-                        return words[words.length - 2].equals("404");
-                    })
-                    .forEach(lines::add);
-        } catch (Exception e) {
-            e.printStackTrace();
+        List<String> lines = null;
+        if (!isBlank(file)) {
+            try (BufferedReader buffer = new BufferedReader(new FileReader(file))) {
+                lines = buffer.lines()
+                        .filter(line -> {
+                            String[] words = line.split(DELIMITER_LINE);
+                            return words[words.length - 2].equals("404");
+                        })
+                        .collect(Collectors.toList());
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
         }
         return lines;
     }
 
     public static void save(List<String> log, String file) {
-        try (PrintWriter out = new PrintWriter(
-                new BufferedOutputStream(
-                        new FileOutputStream(file)
-                ))) {
-            out.println(log);
-        } catch (Exception e) {
-            e.printStackTrace();
+        if (!isBlank(file)) {
+            File targetFile = new File(file);
+            if (targetFile.exists()) {
+                try (FileWriter addInfo = new FileWriter(file, true)) {
+                    for (String line : log) {
+                        addInfo.write(line + System.lineSeparator());
+                    }
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            } else {
+                try (PrintWriter out = new PrintWriter(
+                        new BufferedOutputStream(
+                                new FileOutputStream(file)
+                        ))) {
+                    out.println(log);
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
         }
     }
 
     public static void main(String[] args) {
-        List<String> log = filter("log.txt");
-        save(log, "404.txt");
+        List<String> firstLog = filter("log.txt");
+        List<String> secondLog = filter("log2.txt");
+        save(firstLog, "404.txt");
+        save(secondLog, "404.txt");
     }
 }
